feat: adhoc UX prototype
