+++ client/src/views/Concepts/index.js
@@ -1,20 +1,20 @@
 import React, { useEffect, useState, useRef, useMemo, useCallback } from "react";
 import { createUseStyles } from "react-jss";
 import MainLayout from "../../layouts/MainLayout";
-import searchIcon from "../Home/images/search.svg";
-import caretDown from "../Home/images/caretDown.svg";
+import searchIcon from "../Tasks/images/search.svg";
+import caretDown from "../Tasks/images/caretDown.svg";
 import { CircularProgress } from "@material-ui/core";
 import useOutsideAlerter, {
   conceptStatus,
   Breakpoints,
 } from "../../utils/utils";
 import { getAdminsSimple } from "../../api/usersApi";
-import check from "../Home/images/check.svg";
-import checked from "../Home/images/checked.svg";
+import check from "../Tasks/images/check.svg";
+import checked from "../Tasks/images/checked.svg";
 import { isMobile } from "react-device-detect";
 import {fetchAllConcepts} from "../../api/bountiesApi";
 import ConceptListItem from "../../components/ConceptListItem";
-import addIcon from "../Home/images/add.svg";
+import addIcon from "../Tasks/images/add.svg";
 import RequestNewConceptView from "../RequestNewConcept";
 
 const useStyles = createUseStyles({
