+++ client/src/views/Dashboard/index.js
@@ -1,22 +1,23 @@
-import React, { useState, useEffect } from "react";
+import React, {useState, useEffect, useCallback} from "react";
 import { createUseStyles } from "react-jss";
 import MainLayout from "../../layouts/MainLayout";
 import Lottie from "react-lottie";
 import * as animationData from "./done.json";
 import { fetchDashboard, fetchNotifications } from "../../api/global";
 import NotificationItem from "../../components/NotificationItem";
-import checkedIcon from "../Home/images/checked.svg";
+import checkedIcon from "../Tasks/images/checked.svg";
 import { CircularProgress } from "@material-ui/core";
 import { readAllNotifications } from "../../api/notificationsApi";
-import caretDown from "../Home/images/caretDown.svg";
+import caretDown from "../Tasks/images/caretDown.svg";
 import { truncate, Breakpoints } from "../../utils/utils";
 import { useHistory } from "react-router";
 import cx from "classnames";
 import {
   BountyLocation,
   ConceptLocation,
-  PaymentsLocation,
+  PaymentsLocation, ProfileLocation,
 } from "../../Locations";
+import useGlobalState from "../../state";
 
 const defaultOptions = {
   loop: false,
@@ -35,8 +36,8 @@ const useStyles = createUseStyles({
     marginTop: "32px",
     color: "#0B0F3B",
     display: "flex",
-    justifyContent: "space-between",
-    flexWrap: "wrap",
+    flexDirection: "column",
+    alignItems: "flex-start",
   },
   markAll: {
     marginLeft: "0px",
@@ -46,6 +47,19 @@ const useStyles = createUseStyles({
     color: "white",
     fontSize: "12px",
     marginTop: 0,
+  },
+  viewPofile: {
+    backgroundColor: "#0B0F3B",
+    color: "white",
+    borderRadius: "4px",
+    padding: "8px",
+    display: "flex",
+    alignItems: "center",
+    cursor: "pointer",
+    fontSize: "12px",
+    lineHeight: "15px",
+    fontWeight: 600,
+    userSelect: "none",
   },
   mobileHeading: { display: "block", alignItems: "center" },
   notifBadge: {
@@ -100,8 +114,21 @@ const useStyles = createUseStyles({
     padding: "8px",
     fontSize: "12px",
   },
-  rightColumn: { marginLeft: "0px", flexShrink: 0, marginTop: "24px" },
-  [`@media (min-width: ${Breakpoints.sm}px)`]: {
+  mainContainer: {
+    width: "100%",
+    display: "flex",
+  },
+  rightColumn: {
+    display: "flex",
+    flexDirection: "column",
+    flex: 1,
+  },
+  leftColumn: {
+    marginLeft: 0,
+    flexShrink: 0,
+    marginTop: 24,
+  },
+  [`@media (min-width: ${Breakpoints.lg}px)`]: {
     markAll: {
       marginLeft: "24px",
       cursor: "pointer",
@@ -114,14 +141,11 @@ const useStyles = createUseStyles({
     mobileHeading: { display: "flex", alignItems: "center" },
     rightColumn: { marginLeft: "24px", flexShrink: 0, marginTop: "0px" },
     container: {
-      maxWidth: "1050px",
+      maxWidth: 1600,
       margin: "auto",
       padding: "0 88px 88px 88px",
       marginTop: "32px",
       color: "#0B0F3B",
-      display: "flex",
-      justifyContent: "space-between",
-      flexWrap: "nowrap",
     },
   },
 });
@@ -141,6 +165,7 @@ export default function DashboardView({ match }) {
   const [showPendingAdminApproval, setShowPendingAdminApproval] =
     useState(false);
   const [showUnpaidTasks, setShowUnpaidTasks] = useState(false);
+  const { loggedInUser } = useGlobalState();
   const history = useHistory();
 
   const styles = useStyles();
@@ -169,224 +194,134 @@ export default function DashboardView({ match }) {
       .then((result) => setNotifications(result));
   }, []);
 
+  const handleViewProfile = useCallback(() => {
+    if (loggedInUser) {
+      history.push(ProfileLocation(loggedInUser.username));
+    }
+  }, [loggedInUser, history]);
+
   return (
     <MainLayout match={match}>
       <div className={styles.container}>
-        <div style={{ width: "100%" }}>
-          <div className={styles.header}>DASHBOARD</div>
-          <div style={{ marginTop: "32px", color: "white", fontSize: "18px" }}>
-            {dashboardLoading || !dashboardItems ? (
-              <div style={{ display: "flex", justifyContent: "center" }}>
-                <CircularProgress style={{ color: "white" }} />
-              </div>
-            ) : (dashboardItems.unpaidTasks?.length === 0 ||
-                !dashboardItems.unpaidTasks) &&
-              (dashboardItems.pendingAdminApproval?.length === 0 ||
-                !dashboardItems.pendingAdminApproval) &&
-              (dashboardItems.pendingConcepts?.length === 0 ||
-                !dashboardItems.pendingConcepts) &&
-              dashboardItems.pendingContributorCompletion.length === 0 &&
-              dashboardItems.pendingContributorModify.length === 0 ? (
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                }}
-              >
-                <div>
-                  <Lottie
-                    options={defaultOptions}
-                    height={50}
-                    width={50}
-                    isClickToPauseDisabled={true}
-                  />
-                  <div style={{ marginTop: "4px" }}>You're all caught up </div>
+        {
+          loggedInUser &&
+          <div className={styles.viewPofile} onClick={handleViewProfile}>View Profile</div>
+        }
+        <div className={styles.mainContainer}>
+          <div className={styles.leftColumn}>
+            {notifications?.length > 0 && (
+              <>
+                <div className={cx(styles.header, styles.mobileHeading)}>
+                  <div>
+                    NOTIFICATIONS{" "}
+                    {notifications.filter((notif) => !notif.isRead).length >
+                    0 && (
+                      <span className={styles.notifBadge}>
+                        {notifications.filter((notif) => !notif.isRead).length}
+                      </span>
+                    )}
+                  </div>
+                  {notifications.filter((notif) => !notif.isRead).length > 0 && (
+                    <div
+                      className={cx(styles.header, styles.markAll)}
+                      onClick={() => !loading && setAllRead()}
+                    >
+                      {loading ? (
+                        <div style={{ marginRight: "12px" }}>
+                          <CircularProgress
+                            style={{ color: "white" }}
+                            size={14}
+                          />
+                        </div>
+                      ) : (
+                        <img
+                          src={checkedIcon}
+                          alt="read"
+                          style={{
+                            marginRight: "8px",
+                            width: "14px",
+                            height: "14px",
+                          }}
+                        />
+                      )}
+                      Mark all notifications as read
+                    </div>
+                  )}
                 </div>
-              </div>
-            ) : (
-              <div>
-                {dashboardItems.unpaidTasks?.length > 0 && (
-                  <div className={styles.dashboardItemContainer}>
-                    <div
-                      className={styles.dashboardItemHeader}
-                      onClick={() => setShowUnpaidTasks(!showUnpaidTasks)}
-                    >
-                      <div>
-                        ⚠️ You have{" "}
-                        <span
-                          className={styles.notifBadge}
-                          style={{ margin: "0px 2px" }}
-                        >
-                          {dashboardItems.unpaidTasks.length}
-                        </span>{" "}
-                        unpaid task
-                        {dashboardItems.unpaidTasks.length > 1 && "s"} to payout
-                      </div>
-                      <img
-                        src={caretDown}
-                        alt="dropdown"
-                        style={{
-                          marginLeft: "6px",
-                          transition: "all 0.2s",
-                          transform: showUnpaidTasks
-                            ? "rotate(-180deg)"
-                            : "rotate(0deg)",
-                        }}
-                      />
-                    </div>
-                    {showUnpaidTasks && (
-                      <div>
-                        {dashboardItems.unpaidTasks.map((item, i) => (
-                          <div
-                            style={{
-                              marginTop: "24px",
-                              color: "#0B0F3B",
-                              fontSize: "12px",
-                              fontWeight: "normal",
-                              display: "flex",
-                              alignItems: "center",
-                              cursor: "pointer",
-                            }}
-                            onClick={() => history.push(PaymentsLocation)}
-                          >
-                            <div className={styles.index}>{i + 1}</div>
-                            <div
-                              style={{ textDecoration: "underline" }}
-                              dangerouslySetInnerHTML={{
-                                __html: truncate(item.description, 60),
-                              }}
-                            />
-                          </div>
-                        ))}
-                      </div>
-                    )}
+                <div style={{ marginTop: "32px" }}>
+                  {notifications?.map((notification) => (
+                    <NotificationItem notification={notification} />
+                  ))}
+                </div>
+              </>
+            )}
+          </div>
+          <div className={styles.rightColumn}>
+            <div className={styles.header}>DASHBOARD</div>
+            <div style={{ marginTop: "32px", color: "white", fontSize: "18px" }}>
+              {dashboardLoading || !dashboardItems ? (
+                <div style={{ display: "flex", justifyContent: "center" }}>
+                  <CircularProgress style={{ color: "white" }} />
+                </div>
+              ) : (dashboardItems.unpaidTasks?.length === 0 ||
+                  !dashboardItems.unpaidTasks) &&
+                (dashboardItems.pendingAdminApproval?.length === 0 ||
+                  !dashboardItems.pendingAdminApproval) &&
+                (dashboardItems.pendingConcepts?.length === 0 ||
+                  !dashboardItems.pendingConcepts) &&
+                dashboardItems.pendingContributorCompletion.length === 0 &&
+                dashboardItems.pendingContributorModify.length === 0 ? (
+                <div
+                  style={{
+                    display: "flex",
+                    alignItems: "center",
+                    justifyContent: "center",
+                  }}
+                >
+                  <div>
+                    <Lottie
+                      options={defaultOptions}
+                      height={50}
+                      width={50}
+                      isClickToPauseDisabled={true}
+                    />
+                    <div style={{ marginTop: "4px" }}>You're all caught up </div>
                   </div>
-                )}
-                {dashboardItems.pendingAdminApproval?.length > 0 && (
-                  <div className={styles.dashboardItemContainer}>
-                    <div
-                      className={styles.dashboardItemHeader}
-                      onClick={() =>
-                        setShowPendingAdminApproval(!showPendingAdminApproval)
-                      }
-                    >
-                      <div>
-                        ⚠️ You have{" "}
-                        <span
-                          className={styles.notifBadge}
-                          style={{ margin: "0px 2px" }}
-                        >
-                          {dashboardItems.pendingAdminApproval.length}
-                        </span>{" "}
-                        task completion
-                        {dashboardItems.pendingAdminApproval.length > 1 &&
-                          "s"}{" "}
-                        pending your review
-                      </div>
-                      <img
-                        src={caretDown}
-                        alt="dropdown"
-                        style={{
-                          marginLeft: "6px",
-                          transition: "all 0.2s",
-                          transform: showPendingAdminApproval
-                            ? "rotate(-180deg)"
-                            : "rotate(0deg)",
-                        }}
-                      />
-                    </div>
-                    {showPendingAdminApproval && (
-                      <div>
-                        {dashboardItems.pendingAdminApproval.map((item, i) => (
-                          <div
-                            style={{
-                              marginTop: "24px",
-                              color: "#0B0F3B",
-                              fontSize: "12px",
-                              fontWeight: "normal",
-                              display: "flex",
-                              cursor: "pointer",
-                            }}
-                            onClick={() =>
-                              history.push(
-                                BountyLocation(item.bountyDisplayURL)
-                              )
-                            }
+                </div>
+              ) : (
+                <div>
+                  {dashboardItems.unpaidTasks?.length > 0 && (
+                    <div className={styles.dashboardItemContainer}>
+                      <div
+                        className={styles.dashboardItemHeader}
+                        onClick={() => setShowUnpaidTasks(!showUnpaidTasks)}
+                      >
+                        <div>
+                          ⚠️ You have{" "}
+                          <span
+                            className={styles.notifBadge}
+                            style={{ margin: "0px 2px" }}
                           >
-                            <div className={styles.index}>{i + 1}</div>
-                            <div>
-                              <div>
-                                <b>
-                                  {item.bountyType === "job"
-                                    ? item.completionUser.username
-                                    : item.completions[
-                                        item.completions.length - 1
-                                      ].completionUser.username}
-                                </b>
-                              </div>
-                              <div
-                                className={styles.quote}
-                                dangerouslySetInnerHTML={{
-                                  __html: truncate(
-                                    item.bountyType === "job"
-                                      ? item.completionDescription
-                                      : item.completions[
-                                          item.completions.length - 1
-                                        ].completionDescription,
-                                    60
-                                  ),
-                                }}
-                              />
-                              <div style={{ marginTop: "4px" }}>
-                                Bounty: <u>{item.bountyTitle}</u>
-                              </div>
-                            </div>
-                          </div>
-                        ))}
+                            {dashboardItems.unpaidTasks.length}
+                          </span>{" "}
+                          unpaid task
+                          {dashboardItems.unpaidTasks.length > 1 && "s"} to payout
+                        </div>
+                        <img
+                          src={caretDown}
+                          alt="dropdown"
+                          style={{
+                            marginLeft: "6px",
+                            transition: "all 0.2s",
+                            transform: showUnpaidTasks
+                              ? "rotate(-180deg)"
+                              : "rotate(0deg)",
+                          }}
+                        />
                       </div>
-                    )}
-                  </div>
-                )}
-                {dashboardItems.pendingContributorCompletion?.length > 0 && (
-                  <div className={styles.dashboardItemContainer}>
-                    <div
-                      className={styles.dashboardItemHeader}
-                      onClick={() =>
-                        setShowPendingContributorCompletion(
-                          !showPendingContributorCompletion
-                        )
-                      }
-                    >
-                      <div>
-                        ⚠️ You have{" "}
-                        <span
-                          className={styles.notifBadge}
-                          style={{ margin: "0px 2px" }}
-                        >
-                          {dashboardItems.pendingContributorCompletion.length}
-                        </span>{" "}
-                        task
-                        {dashboardItems.pendingContributorCompletion.length >
-                          1 && "s"}{" "}
-                        awaiting your completion
-                      </div>
-                      <img
-                        src={caretDown}
-                        alt="dropdown"
-                        style={{
-                          marginLeft: "6px",
-                          transition: "all 0.2s",
-                          transform: showPendingContributorCompletion
-                            ? "rotate(-180deg)"
-                            : "rotate(0deg)",
-                        }}
-                      />
-                    </div>
-                    {showPendingContributorCompletion && (
-                      <div>
-                        {dashboardItems.pendingContributorCompletion.map(
-                          (item, i) => (
+                      {showUnpaidTasks && (
+                        <div>
+                          {dashboardItems.unpaidTasks.map((item, i) => (
                             <div
                               style={{
                                 marginTop: "24px",
@@ -397,11 +332,7 @@ export default function DashboardView({ match }) {
                                 alignItems: "center",
                                 cursor: "pointer",
                               }}
-                              onClick={() =>
-                                history.push(
-                                  BountyLocation(item.bountyDisplayURL)
-                                )
-                              }
+                              onClick={() => history.push(PaymentsLocation)}
                             >
                               <div className={styles.index}>{i + 1}</div>
                               <div
@@ -411,51 +342,47 @@ export default function DashboardView({ match }) {
                                 }}
                               />
                             </div>
-                          )
-                        )}
-                      </div>
-                    )}
-                  </div>
-                )}
-                {dashboardItems.pendingContributorModify?.length > 0 && (
-                  <div className={styles.dashboardItemContainer}>
-                    <div
-                      className={styles.dashboardItemHeader}
-                      onClick={() =>
-                        setShowPendingContributorModify(
-                          !showPendingContributorModify
-                        )
-                      }
-                    >
-                      <div>
-                        ⚠️ You have{" "}
-                        <span
-                          className={styles.notifBadge}
-                          style={{ margin: "0px 2px" }}
-                        >
-                          {dashboardItems.pendingContributorModify.length}
-                        </span>{" "}
-                        task
-                        {dashboardItems.pendingContributorModify.length > 1 &&
-                          "s"}{" "}
-                        awaiting your modifications
-                      </div>
-                      <img
-                        src={caretDown}
-                        alt="dropdown"
-                        style={{
-                          marginLeft: "6px",
-                          transition: "all 0.2s",
-                          transform: showPendingContributorModify
-                            ? "rotate(-180deg)"
-                            : "rotate(0deg)",
-                        }}
-                      />
+                          ))}
+                        </div>
+                      )}
                     </div>
-                    {showPendingContributorModify && (
-                      <div>
-                        {dashboardItems.pendingContributorModify.map(
-                          (item, i) => (
+                  )}
+                  {dashboardItems.pendingAdminApproval?.length > 0 && (
+                    <div className={styles.dashboardItemContainer}>
+                      <div
+                        className={styles.dashboardItemHeader}
+                        onClick={() =>
+                          setShowPendingAdminApproval(!showPendingAdminApproval)
+                        }
+                      >
+                        <div>
+                          ⚠️ You have{" "}
+                          <span
+                            className={styles.notifBadge}
+                            style={{ margin: "0px 2px" }}
+                          >
+                            {dashboardItems.pendingAdminApproval.length}
+                          </span>{" "}
+                          task completion
+                          {dashboardItems.pendingAdminApproval.length > 1 &&
+                            "s"}{" "}
+                          pending your review
+                        </div>
+                        <img
+                          src={caretDown}
+                          alt="dropdown"
+                          style={{
+                            marginLeft: "6px",
+                            transition: "all 0.2s",
+                            transform: showPendingAdminApproval
+                              ? "rotate(-180deg)"
+                              : "rotate(0deg)",
+                          }}
+                        />
+                      </div>
+                      {showPendingAdminApproval && (
+                        <div>
+                          {dashboardItems.pendingAdminApproval.map((item, i) => (
                             <div
                               style={{
                                 marginTop: "24px",
@@ -475,18 +402,22 @@ export default function DashboardView({ match }) {
                               <div>
                                 <div>
                                   <b>
-                                    {
-                                      item.reviews[item.reviews.length - 1]
-                                        .reviewUser.username
-                                    }
+                                    {item.bountyType === "job"
+                                      ? item.completionUser.username
+                                      : item.completions[
+                                          item.completions.length - 1
+                                        ].completionUser.username}
                                   </b>
                                 </div>
                                 <div
                                   className={styles.quote}
                                   dangerouslySetInnerHTML={{
                                     __html: truncate(
-                                      item.reviews[item.reviews.length - 1]
-                                        .reviewComments,
+                                      item.bountyType === "job"
+                                        ? item.completionDescription
+                                        : item.completions[
+                                            item.completions.length - 1
+                                          ].completionDescription,
                                       60
                                     ),
                                   }}
@@ -496,125 +427,231 @@ export default function DashboardView({ match }) {
                                 </div>
                               </div>
                             </div>
-                          )
-                        )}
-                      </div>
-                    )}
-                  </div>
-                )}
-                {dashboardItems.pendingConcepts?.length > 0 && (
-                  <div className={styles.dashboardItemContainer}>
-                    <div
-                      className={styles.dashboardItemHeader}
-                      onClick={() =>
-                        setShowPendingConcepts(!showPendingConcepts)
-                      }
-                    >
-                      <div>
-                        ⚠️ There{" "}
-                        {dashboardItems.pendingConcepts.length > 1
-                          ? "are"
-                          : "is"}{" "}
-                        <span
-                          className={styles.notifBadge}
-                          style={{ margin: "0px 2px" }}
-                        >
-                          {dashboardItems.pendingConcepts.length}
-                        </span>{" "}
-                        concept
-                        {dashboardItems.pendingConcepts.length > 1 && "s"}{" "}
-                        pending approval
-                      </div>
-                      <img
-                        src={caretDown}
-                        alt="dropdown"
-                        style={{
-                          marginLeft: "6px",
-                          transition: "all 0.2s",
-                          transform: showPendingConcepts
-                            ? "rotate(-180deg)"
-                            : "rotate(0deg)",
-                        }}
-                      />
+                          ))}
+                        </div>
+                      )}
                     </div>
-                    {showPendingConcepts && (
-                      <div>
-                        {dashboardItems.pendingConcepts.map((item, i) => (
-                          <div
-                            style={{
-                              marginTop: "24px",
-                              color: "#0B0F3B",
-                              fontSize: "12px",
-                              fontWeight: "normal",
-                              display: "flex",
-                              alignItems: "center",
-                              cursor: "pointer",
-                            }}
-                            onClick={() =>
-                              history.push(ConceptLocation(item.displayURL))
-                            }
-                          >
-                            <div className={styles.index}>{i + 1}</div>
-                            <div style={{ textDecoration: "underline" }}>
-                              {item.title}
-                            </div>
-                          </div>
-                        ))}
-                      </div>
-                    )}
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-        <div className={styles.rightColumn}>
-          {notifications?.length > 0 && (
-            <>
-              <div className={cx(styles.header, styles.mobileHeading)}>
-                <div>
-                  NOTIFICATIONS{" "}
-                  {notifications.filter((notif) => !notif.isRead).length >
-                    0 && (
-                    <span className={styles.notifBadge}>
-                      {notifications.filter((notif) => !notif.isRead).length}
-                    </span>
                   )}
-                </div>
-                {notifications.filter((notif) => !notif.isRead).length > 0 && (
-                  <div
-                    className={cx(styles.header, styles.markAll)}
-                    onClick={() => !loading && setAllRead()}
-                  >
-                    {loading ? (
-                      <div style={{ marginRight: "12px" }}>
-                        <CircularProgress
-                          style={{ color: "white" }}
-                          size={14}
+                  {dashboardItems.pendingContributorCompletion?.length > 0 && (
+                    <div className={styles.dashboardItemContainer}>
+                      <div
+                        className={styles.dashboardItemHeader}
+                        onClick={() =>
+                          setShowPendingContributorCompletion(
+                            !showPendingContributorCompletion
+                          )
+                        }
+                      >
+                        <div>
+                          ⚠️ You have{" "}
+                          <span
+                            className={styles.notifBadge}
+                            style={{ margin: "0px 2px" }}
+                          >
+                            {dashboardItems.pendingContributorCompletion.length}
+                          </span>{" "}
+                          task
+                          {dashboardItems.pendingContributorCompletion.length >
+                            1 && "s"}{" "}
+                          awaiting your completion
+                        </div>
+                        <img
+                          src={caretDown}
+                          alt="dropdown"
+                          style={{
+                            marginLeft: "6px",
+                            transition: "all 0.2s",
+                            transform: showPendingContributorCompletion
+                              ? "rotate(-180deg)"
+                              : "rotate(0deg)",
+                          }}
                         />
                       </div>
-                    ) : (
-                      <img
-                        src={checkedIcon}
-                        alt="read"
-                        style={{
-                          marginRight: "8px",
-                          width: "14px",
-                          height: "14px",
-                        }}
-                      />
-                    )}
-                    Mark all notifications as read
-                  </div>
-                )}
-              </div>
-              <div style={{ marginTop: "32px" }}>
-                {notifications?.map((notification) => (
-                  <NotificationItem notification={notification} />
-                ))}
-              </div>
-            </>
-          )}
+                      {showPendingContributorCompletion && (
+                        <div>
+                          {dashboardItems.pendingContributorCompletion.map(
+                            (item, i) => (
+                              <div
+                                style={{
+                                  marginTop: "24px",
+                                  color: "#0B0F3B",
+                                  fontSize: "12px",
+                                  fontWeight: "normal",
+                                  display: "flex",
+                                  alignItems: "center",
+                                  cursor: "pointer",
+                                }}
+                                onClick={() =>
+                                  history.push(
+                                    BountyLocation(item.bountyDisplayURL)
+                                  )
+                                }
+                              >
+                                <div className={styles.index}>{i + 1}</div>
+                                <div
+                                  style={{ textDecoration: "underline" }}
+                                  dangerouslySetInnerHTML={{
+                                    __html: truncate(item.description, 60),
+                                  }}
+                                />
+                              </div>
+                            )
+                          )}
+                        </div>
+                      )}
+                    </div>
+                  )}
+                  {dashboardItems.pendingContributorModify?.length > 0 && (
+                    <div className={styles.dashboardItemContainer}>
+                      <div
+                        className={styles.dashboardItemHeader}
+                        onClick={() =>
+                          setShowPendingContributorModify(
+                            !showPendingContributorModify
+                          )
+                        }
+                      >
+                        <div>
+                          ⚠️ You have{" "}
+                          <span
+                            className={styles.notifBadge}
+                            style={{ margin: "0px 2px" }}
+                          >
+                            {dashboardItems.pendingContributorModify.length}
+                          </span>{" "}
+                          task
+                          {dashboardItems.pendingContributorModify.length > 1 &&
+                            "s"}{" "}
+                          awaiting your modifications
+                        </div>
+                        <img
+                          src={caretDown}
+                          alt="dropdown"
+                          style={{
+                            marginLeft: "6px",
+                            transition: "all 0.2s",
+                            transform: showPendingContributorModify
+                              ? "rotate(-180deg)"
+                              : "rotate(0deg)",
+                          }}
+                        />
+                      </div>
+                      {showPendingContributorModify && (
+                        <div>
+                          {dashboardItems.pendingContributorModify.map(
+                            (item, i) => (
+                              <div
+                                style={{
+                                  marginTop: "24px",
+                                  color: "#0B0F3B",
+                                  fontSize: "12px",
+                                  fontWeight: "normal",
+                                  display: "flex",
+                                  cursor: "pointer",
+                                }}
+                                onClick={() =>
+                                  history.push(
+                                    BountyLocation(item.bountyDisplayURL)
+                                  )
+                                }
+                              >
+                                <div className={styles.index}>{i + 1}</div>
+                                <div>
+                                  <div>
+                                    <b>
+                                      {
+                                        item.reviews[item.reviews.length - 1]
+                                          .reviewUser.username
+                                      }
+                                    </b>
+                                  </div>
+                                  <div
+                                    className={styles.quote}
+                                    dangerouslySetInnerHTML={{
+                                      __html: truncate(
+                                        item.reviews[item.reviews.length - 1]
+                                          .reviewComments,
+                                        60
+                                      ),
+                                    }}
+                                  />
+                                  <div style={{ marginTop: "4px" }}>
+                                    Bounty: <u>{item.bountyTitle}</u>
+                                  </div>
+                                </div>
+                              </div>
+                            )
+                          )}
+                        </div>
+                      )}
+                    </div>
+                  )}
+                  {dashboardItems.pendingConcepts?.length > 0 && (
+                    <div className={styles.dashboardItemContainer}>
+                      <div
+                        className={styles.dashboardItemHeader}
+                        onClick={() =>
+                          setShowPendingConcepts(!showPendingConcepts)
+                        }
+                      >
+                        <div>
+                          ⚠️ There{" "}
+                          {dashboardItems.pendingConcepts.length > 1
+                            ? "are"
+                            : "is"}{" "}
+                          <span
+                            className={styles.notifBadge}
+                            style={{ margin: "0px 2px" }}
+                          >
+                            {dashboardItems.pendingConcepts.length}
+                          </span>{" "}
+                          concept
+                          {dashboardItems.pendingConcepts.length > 1 && "s"}{" "}
+                          pending approval
+                        </div>
+                        <img
+                          src={caretDown}
+                          alt="dropdown"
+                          style={{
+                            marginLeft: "6px",
+                            transition: "all 0.2s",
+                            transform: showPendingConcepts
+                              ? "rotate(-180deg)"
+                              : "rotate(0deg)",
+                          }}
+                        />
+                      </div>
+                      {showPendingConcepts && (
+                        <div>
+                          {dashboardItems.pendingConcepts.map((item, i) => (
+                            <div
+                              style={{
+                                marginTop: "24px",
+                                color: "#0B0F3B",
+                                fontSize: "12px",
+                                fontWeight: "normal",
+                                display: "flex",
+                                alignItems: "center",
+                                cursor: "pointer",
+                              }}
+                              onClick={() =>
+                                history.push(ConceptLocation(item.displayURL))
+                              }
+                            >
+                              <div className={styles.index}>{i + 1}</div>
+                              <div style={{ textDecoration: "underline" }}>
+                                {item.title}
+                              </div>
+                            </div>
+                          ))}
+                        </div>
+                      )}
+                    </div>
+                  )}
+                </div>
+              )}
+            </div>
+          </div>
         </div>
       </div>
     </MainLayout>
